+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added `VirtuosoApacheHttpClientRdfStoreServiceTest` tests
 - removing deprecated `CognizoneException` class and `SparqlService.queryForModel` and `SparqlService.upload` methods
 - adding `SparqlService.isEmptyGraph` method
+- parameterization of integration tests for Virtuoso and GraphDB
 
 1.0.0
 - extracted from ASquare